Driver!!
Driver that has both single hash table and hash table/BST combo.  Times both methods at the end... that part not quite working yet. I had to change the header files of BST and hash to be able to do the single hash table and the timer so Jocob's main wont work with those headers.